requirements.txt
pillow
nicegui